Develop

Closes ICRC-21

See merge request origyn/engineering/opensource/claimlink!24

+++ backend/canisters/claimlink/impl/src/queries/estimate_mint_cost.rs
@@ -11,16 +11,14 @@ pub use claimlink_api::queries::estimate_mint_cost::{
 pub fn estimate_mint_cost(args: EstimateMintCostArgs) -> EstimateMintCostResponse {
     let (mint_pricing, usd_per_ogy_e8s) = read_state(|s| {
         (
-            s.data.mint_pricing, // Copy
+            s.data.mint_pricing,
             s.data.ogy_price.map(|p| p.usd_per_ogy_e8s),
         )
     });
 
-    let usd_per_ogy_e8s =
-        usd_per_ogy_e8s.ok_or(EstimateMintCostError::OgyPriceNotAvailable)?;
+    let usd_per_ogy_e8s = usd_per_ogy_e8s.ok_or(EstimateMintCostError::OgyPriceNotAvailable)?;
 
-    let mint_pricing =
-        mint_pricing.ok_or(EstimateMintCostError::MintPricingNotConfigured)?;
+    let mint_pricing = mint_pricing.ok_or(EstimateMintCostError::MintPricingNotConfigured)?;
 
     let total_file_size_bytes: u64 = args.total_file_size_bytes.0.try_into().unwrap_or(u64::MAX);
 
@@ -32,7 +30,7 @@ pub fn estimate_mint_cost(args: EstimateMintCostArgs) -> EstimateMintCostRespons
         usd_per_ogy_e8s,
     );
 
-    let base_fee_usd_e8s = (mint_pricing.base_mint_fee_usd_e8s as u64) * args.num_mints;
+    let base_fee_usd_e8s = mint_pricing.base_mint_fee_usd_e8s * args.num_mints;
     let storage_fee_usd_e8s = total_usd_e8s.saturating_sub(base_fee_usd_e8s);
 
     Ok(MintCostEstimate {
